requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.0
arxiv==1.4.3
async-timeout==4.0.3
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==2.0.12
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.115.0
feedparser==6.0.11
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
google-ai-generativelanguage==0.1.0
google-api-core==2.20.0
google-auth==2.35.0
google-generativeai==0.1.0rc1
googleapis-common-protos==1.65.0
gradio==3.23.0
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
importlib-resources==6.4.5
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
linkify-it-py==2.0.3
logging==0.4.9.6
Markdown==3.3.4
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.1.0
narwhals==1.8.3
numpy==1.24.4
orjson==3.10.7
packaging==24.1
pandas==2.0.3
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
proto-plus==1.24.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.9.2
pydantic-core==2.23.4
pydub==0.25.1
PyMuPDF==1.21.1
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5-sip==12.15.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.15
python-dateutil==2.9.0.post0
python-multipart==0.0.10
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.26.0
rpds-py==0.20.0
rsa==4.9
semantic-version==2.10.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
tenacity==8.2.2
tiktoken==0.3.2
tqdm==4.66.5
typing-extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==1.26.20
uvicorn==0.30.6
websockets==13.1
yarl==1.13.0
zipp==3.20.2
